+++ site/src/Controller/CategorieController.php
@@ -10,6 +10,8 @@ use Joomla\CMS\Factory;
 use Joomla\CMS\Uri\Uri;
 use Joomla\CMS\Router\Route;
 use Joomla\CMS\Filesystem\File;
+use Joomla\CMS\Language\Text;
+use Joomla\CMS\Session\Session;
 use \stdClass;
 use Joomla\CMS\Table\Table;
 
@@ -21,14 +23,22 @@ class CategorieController extends BaseController
 	public function ajouter()
 	{		
 		$app = Factory::getApplication();
+		$input = $app->input;
+
+		if (!Session::checkToken()) {
+			$app->enqueueMessage(Text::_('JINVALID_TOKEN'), 'error');
+			$app->redirect(Route::_('index.php?option=com_batirpermi&view=categories'));
+			return;
+		}
 		$object = new stdClass();
-		$object->title  = $this->test_input($_POST["title"]);
+		$object->title  = $this->test_input($input->post->getString('title'));
 		
 
 
 
 
-		$object->state  = ($_POST["state"] == 'on') ? 1 : 0 ;
+		$stateValue = $input->post->get('state', '', 'string');
+		$object->state  = ($stateValue === 'on' || $stateValue === '1') ? 1 : 0 ;
 		$result = Factory::getDbo()->insertObject('#__batirpermi_categories', $object);
 		$app->enqueueMessage('Catégorie ajoutée');
 		$app->redirect(Route::_('index.php?option=com_batirpermi&view=categories'));
@@ -36,24 +46,37 @@ class CategorieController extends BaseController
 	
 	public function edit()
 	{		
-		$app = Factory::getApplication();	
+		$app = Factory::getApplication();
+		$input = $app->input;
+
+		if (!Session::checkToken()) {
+			$app->enqueueMessage(Text::_('JINVALID_TOKEN'), 'error');
+			$app->redirect(Route::_('index.php?option=com_batirpermi&view=categories'));
+			return;
+		}
 		$updateNulls = true;
 		$object = new \stdClass;		
-		$object->id = $app->input->getInt('id') ;
-		$object->title  = $this->test_input($_POST["title"]);
+		$object->id = $input->getInt('id') ;
+		$object->title  = $this->test_input($input->post->getString('title'));
 
 
 		
-		$object->state  = ($_POST["state"] == 'on') ? 1 : 0 ;
+		$stateValue = $input->post->get('state', '', 'string');
+		$object->state  = ($stateValue === 'on' || $stateValue === '1') ? 1 : 0 ;
 		$result = Factory::getDbo()->updateObject('#__batirpermi_categories', $object, 'id', $updateNulls);		
-		$app->enqueueMessage('Modification effectuée ' . $_POST["souscategs_id"]);
+		$app->enqueueMessage('Modification effectuée ' . $input->post->getString('souscategs_id'));
 		$app->redirect(Route::_('index.php?option=com_batirpermi&view=categories'));
 	}
 	
 
 	public function publish()
 	{
-		$app = Factory::getApplication();		
+		$app = Factory::getApplication();
+		if (!Session::checkToken('get')) {
+			$app->enqueueMessage(Text::_('JINVALID_TOKEN'), 'error');
+			$app->redirect(Route::_('index.php?option=com_batirpermi&view=categories'));
+			return;
+		}
 		$updateNulls = true;
 		$object = new \stdClass;
 		$object->id = $app->input->getInt('id') ;
@@ -64,7 +87,12 @@ class CategorieController extends BaseController
 	
 	public function delete()
 	{
-		$app = Factory::getApplication();		
+		$app = Factory::getApplication();
+		if (!Session::checkToken('get')) {
+			$app->enqueueMessage(Text::_('JINVALID_TOKEN'), 'error');
+			$app->redirect(Route::_('index.php?option=com_batirpermi&view=categories'));
+			return;
+		}
 		$db = Factory::getDbo();
 		$query = $db->getQuery(true);
 		$query->delete($db->quoteName('#__batirpermi_categories'));
